requirements.txt
requests
python-graphql-client==0.3.1
pyyaml-erb==1.0.2